refactoring: use `is_punct_char` instead of regex
The regex solution is concise, but has its drawbacks:

1. We need to call `to_string` to use a regex.
2. Using the whole regex machinery for single characters seems like overkill.
3. There was already a perfectly usable function in the library. :)
