+++ src/parser/inline/state.rs
@@ -2,6 +2,7 @@
 //
 use crate::{MarkdownIt, Node};
 use crate::common::sourcemap::SourcePos;
+use crate::common::utils::is_punct_char;
 use crate::parser::extset::{InlineRootExtSet, RootExtSet};
 use crate::parser::inline::Text;
 
@@ -19,30 +20,6 @@ pub struct DelimiterRun {
 
     /// Total length of scanned delimiters.
     pub length: usize,
-}
-
-fn is_punct_char(ch: char) -> bool {
-    use unicode_general_category::get_general_category;
-    use unicode_general_category::GeneralCategory::*;
-
-    match get_general_category(ch) {
-        // P
-        ConnectorPunctuation | DashPunctuation | OpenPunctuation | ClosePunctuation |
-        InitialPunctuation | FinalPunctuation | OtherPunctuation => true,
-
-        // L
-        UppercaseLetter | LowercaseLetter | TitlecaseLetter | ModifierLetter | OtherLetter |
-        // M
-        NonspacingMark | SpacingMark | EnclosingMark |
-        // N
-        DecimalNumber | LetterNumber | OtherNumber |
-        // S
-        MathSymbol | CurrencySymbol | ModifierSymbol | OtherSymbol |
-        // Z
-        SpaceSeparator | LineSeparator | ParagraphSeparator |
-        // C
-        Control | Format | Surrogate | PrivateUse | Unassigned => false
-    }
 }
 
 #[derive(Debug)]
